benchmarks/dockerize/zero/requirements.txt
# zeroapi
pyzmq
PyJWT
aioredis
uvloop
msgpack
aiohttp
gunicorn
redis
sanic